Bring code and .eslintrc in line
The .eslintrc rules did not match the style of the code. This change applies
the existing rules for one-off cases, and changes the rules to match the
code in cases where the code largely deviates from them anyway. In
particular, in line with "react/jsx-no-target-blank", it adds "noopener
noreferrer" as necessary (see:
https://mathiasbynens.github.io/rel-noopener/).

.eslintrc is also formatted for readability. Of note, each instance of
`[0]` is changed to `["off"]` for clarity of purpose.